feat(mac): implement macOS installer with launcher and Info.plist
