+++ npx-cli/src/mac/shim.d.ts
@@ -1,0 +1,9 @@
+declare module '*.sh' {
+  const content: string;
+  export default content;
+}
+
+declare module '*.plist' {
+  const content: string;
+  export default content;
+}
